Add MIT License to the project
A MIT license file has been added to the project repository. This license grants permission to any user to freely use, modify, and distribute the software. It sets the conditions for using the software, including the need to include the copyright notice and permission notice in all copies or significant parts of the software.